+++ app/provision.sh
@@ -25,7 +25,72 @@ function statuses() {
     echo "http://${2}"
 }
 
+function setSugarBashAlias(){
+    cd /home/vagrant
+    echo "alias helper='bash /home/vagrant/git-repo/sugarBash/helper.sh'" >> ./.bashrc
+}
+
+function resetDeployer() {
+
+    npm install -g grunt-cli grunt bower
+    cd /var/www/html/deployer/
+    git init
+    git remote add origin git@github.com:svnvcristea/stack-deployer.git
+    git remote add upstream git@github.com:sugarcrm/stack-deployer.git
+    git fetch
+    git checkout -b develop
+    git checkout develop
+    git reset --hard origin/develop
+    bower install
+    npm install
+    npm install grunt-cli grunt
+    chown vagrant:vagrant -R ./
+    configDeployer $@
+}
+
+function configDeployer() {
+
+    if [ $3 == "true" ]; then
+        cat > _config.json <<EOL
+{
+  "host":  "localhost",
+  "vagrant": {
+    "host": "localhost",
+    "port": 2222,
+    "username": "vagrant"
+  },
+  "port": {
+    "socket_io": "8333",
+    "node_app": "8334",
+    "web_app": "8080"
+  }
+}
+
+EOL
+    else
+        cat > _config.json <<EOL
+{
+  "host":  "${2}",
+  "vagrant": {
+    "host": "${2}",
+    "port": 22,
+    "username": "vagrant"
+  },
+  "port": {
+    "socket_io": "3333",
+    "node_app": "3334",
+    "web_app": "80"
+  }
+}
+
+EOL
+    fi
+    grunt init
+}
+
 echo "Stack provision: ${1}"
 installTools
 nfsExports
+setSugarBashAlias
+resetDeployer $@
 statuses $1 $2